requirements.txt
opencv-contrib-python
numpy
matplotlib
tensorflow
keras
flask
eventlet
python-socketio==4.6.1
python-engineio==3.13.2
imgaug
scikit-learn